@@ -41,74 +41,38 @@
   "devDependencies": {
     "@babel/core": "^7.17.10",
     "@babel/eslint-parser": "^7.17.0",
-    "@babel/plugin-transform-runtime": "^7.17.10",
-    "@babel/preset-env": "^7.17.10",
-    "@babel/preset-react": "^7.16.7",
-    "@commitlint/cli": "^16.2.4",
+    "@commitlint/cli": "^16.3.0",
     "@commitlint/config-conventional": "^16.2.4",
-    "@fancyapps/ui": "^4.0.26",
-    "@rollup/plugin-babel": "^5.3.1",
-    "@rollup/plugin-commonjs": "^22.0.0",
-    "@rollup/plugin-node-resolve": "^13.3.0",
-    "@rollup/plugin-replace": "^4.0.0",
-    "@types/inquirer": "^8.2.1",
-    "@types/jest": "^27.5.0",
-    "@types/node": "^17.0.31",
-    "@types/semver": "^7.3.9",
-    "@typescript-eslint/eslint-plugin": "^5.22.0",
-    "@typescript-eslint/parser": "^5.22.0",
+    "@typescript-eslint/eslint-plugin": "^5.23.0",
+    "@typescript-eslint/parser": "^5.23.0",
     "@vue/eslint-config-prettier": "7.0.0",
     "@vue/eslint-config-typescript": "10.0.0",
     "apidoc": "^0.51.1",
-    "babel-jest": "^28.1.0",
-    "babel-loader": "^8.2.5",
+    "c8": "^7.11.2",
     "commitizen": "^4.2.4",
     "conventional-changelog-cli": "^2.2.2",
     "cross-env": "^7.0.3",
-    "css-loader": "^6.7.1",
     "cz-conventional-changelog": "^3.3.0",
-    "eslint": "^8.14.0",
+    "eslint": "^8.15.0",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-prettier": "^4.0.0",
     "eslint-plugin-react": "^7.29.4",
     "eslint-plugin-vue": "^8.7.1",
-    "file-loader": "^6.2.0",
-    "html-webpack-plugin": "^5.5.0",
-    "husky": "^7.0.4",
-    "inquirer": "^8.2.4",
-    "jest": "^28.1.0",
-    "lerna": "^4.0.0",
+    "husky": "^8.0.1",
     "lint-staged": "^12.4.1",
-    "marked": "^4.0.15",
-    "postcss-loader": "^6.2.1",
+    "postcss": "^8.4.13",
     "prettier": "^2.6.2",
-    "rollup": "^2.72.0",
-    "rollup-plugin-dts": "^4.2.1",
-    "rollup-plugin-styles": "4.0.0",
-    "rollup-plugin-terser": "^7.0.2",
-    "rollup-plugin-typescript2": "^0.31.2",
-    "rollup-plugin-vue": "^6.0.0",
     "sass": "^1.51.0",
-    "sass-loader": "^12.6.0",
     "sort-package-json": "^1.57.0",
-    "style-loader": "^3.3.1",
     "stylelint": "^14.8.2",
     "stylelint-config-prettier": "^9.0.3",
     "stylelint-config-standard": "^25.0.0",
     "stylelint-config-standard-scss": "^3.0.0",
     "stylelint-order": "^5.0.0",
-    "ts-jest": "^28.0.1",
-    "ts-loader": "^9.3.0",
-    "ts-node": "^10.7.0",
     "typescript": "^4.6.4",
-    "url-loader": "^4.1.1",
-    "vercel": "^24.2.0",
-    "vue-loader": "^17.0.0",
-    "webpack": "^5.72.0",
-    "webpack-bundle-analyzer": "^4.5.0",
-    "webpack-cli": "^4.9.2",
-    "webpack-dev-server": "^4.9.0",
-    "webpack-merge": "^5.8.0"
+    "vercel": "^24.2.3",
+    "vite": "~2.9.9",
+    "vitest": "^0.12.6"
   },
   "apidoc": {
     "title": "Waline API Documentation"
