@@ -1,6 +1,6 @@
 {
   "name": "@waline/client",
-  "version": "2.3.2",
+  "version": "2.5.0",
   "description": "client for waline comment system",
   "keywords": [
     "valine",
@@ -86,12 +86,27 @@
     "vue": "^3.2.33"
   },
   "devDependencies": {
+    "@babel/core": "^7.17.10",
+    "@babel/preset-env": "^7.17.10",
+    "@rollup/plugin-babel": "^5.3.1",
+    "@rollup/plugin-commonjs": "^22.0.0",
+    "@rollup/plugin-node-resolve": "^13.3.0",
+    "@rollup/plugin-replace": "^4.0.0",
     "@types/autosize": "^4.0.1",
     "@types/marked": "^4.0.3",
-    "@vitejs/plugin-vue": "^2.3.2",
-    "vite": "^2.9.8"
+    "@types/node": "^17.0.33",
+    "@vitejs/plugin-vue": "^2.3.3",
+    "@vue/compiler-sfc": "^3.2.33",
+    "@yeger/vue-masonry-wall": "^3.0.31",
+    "rimraf": "^3.0.2",
+    "rollup": "^2.73.0",
+    "rollup-plugin-dts": "^4.2.1",
+    "rollup-plugin-terser": "^7.0.2",
+    "rollup-plugin-ts": "^2.0.7",
+    "typescript": "^4.6.4",
+    "vite": "^2.9.9"
   },
   "engines": {
-    "node": ">=12.20.0"
+    "node": ">=14"
   }
 }
